# apps/api/src/workloads/workloads.controller.ts
import{Controller,Get,Param,Query,Req,UseGuards}from'@nestjs/common';import{ApiTags}from'@nestjs/swagger';import{AuthGuard}from'../auth/auth.guard';import type{CserRequest}from'../common/request-context';import{WorkloadsService}from'./workloads.service';
@ApiTags('workloads')@UseGuards(AuthGuard)@Controller({path:'workloads',version:'1'})export class WorkloadsController{constructor(private readonly service:WorkloadsService){}@Get()list(@Req()req:CserRequest,@Query()query:Record<string,string|undefined>){return this.service.list(req.auth!,query)}@Get(':id')detail(@Req()req:CserRequest,@Param('id')id:string){return this.service.detail(req.auth!,id)}}