# legacy/v2-reference/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSER Workspace Candidate MVP API",
    "version": "2.0.0",
    "description": "Independent candidate concept using fictional cloud-security data. Not an official ESET API or product."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/workloads": {"get":{"tags":["Workloads"],"parameters":[{"name":"page","in":"query","schema":{"type":"integer","minimum":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":10,"maximum":100}},{"name":"provider","in":"query","schema":{"enum":["AZURE","AWS","GCP"]}},{"name":"protectionStatus","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Paginated workloads"}}}},
    "/api/workloads/{id}": {"get":{"tags":["Workloads"],"parameters":[{"$ref":"#/components/parameters/Id"}],"responses":{"200":{"description":"Workload detail"},"404":{"$ref":"#/components/responses/Problem"}}}},
...
  },
  "components": {
    "parameters": {
      "Id": {"name":"id","in":"path","required":true,"schema":{"type":"string"}},
...
    },
    "schemas": {
      "Problem": {"type":"object","required":["title","status","code","detail","correlationId"],"properties":{"type":{"type":"string"},"title":{"type":"string"},"status":{"type":"integer"},"code":{"type":"string"},"detail":{"type":"string"},"correlationId":{"type":"string"},"fieldErrors":{"type":"array","items":{"type":"object"}}}}
    },
    "responses": {"Problem":{"description":"Structured problem response","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}}}
  }
}